Files for creating a Multi-Namespace Array Mock
These files update the ArrayMockLoad.py program. The update supports creating either a single namespace Mock or a two namespace mock. A two namespace mock requires that the instance mofs be split into 2 files. The MockSNIA_ArrayDevInstances.mof file are the instances that will get loaded in the user (e.g., device) namespace. The MockSNIA_ArrayIntInstances.mof file are the instances to be loaded in the interop namespace.